+++ bson/json_util.py
@@ -14,14 +14,15 @@
 
 """Tools for using Python's :mod:`json` module with BSON documents.
 
-This module provides two helper methods `dumps` and `loads` that wrap the native
-:mod:`json` methods and provide explicit BSON conversion to and from
-JSON. :class:`~bson.json_util.JSONOptions` provides a way to control how JSON is
-emitted and parsed, with the default being the legacy PyMongo format.
-:mod:`~bson.json_util` can also generate and parse `canonical extended JSON`_ when
-:data:`~bson.json_util.CANONICAL_JSON_OPTIONS` is provided.
+This module provides two helper methods `dumps` and `loads` that wrap the
+native :mod:`json` methods and provide explicit BSON conversion to and from
+JSON. :class:`~bson.json_util.JSONOptions` provides a way to control how JSON
+is emitted and parsed, with the default being the legacy PyMongo format.
+:mod:`~bson.json_util` can also generate Canonical or Relaxed `Extended JSON`_
+when :const:`CANONICAL_JSON_OPTIONS` or :const:`RELAXED_JSON_OPTIONS` is
+provided, respectively.
 
-.. _canonical extended JSON: https://github.com/mongodb/specifications/blob/master/source/extended-json.rst
+.. _Extended JSON: https://github.com/mongodb/specifications/blob/master/source/extended-json.rst
 
 Example usage (serialization):
 
@@ -40,8 +41,8 @@ Example usage (deserialization):
 .. doctest::
 
    >>> from bson.json_util import loads
-   >>> loads('[{"foo": [1, 2]}, {"bar": {"hello": "world"}}, {"code": {"$scope": {}, "$code": "function x() { return 1; }"}}, {"bin": {"$type": "00", "$binary": "AQIDBA=="}}]')
-   [{u'foo': [1, 2]}, {u'bar': {u'hello': u'world'}}, {u'code': Code('function x() { return 1; }', {})}, {u'bin': Binary('...', 0)}]
+   >>> loads('[{"foo": [1, 2]}, {"bar": {"hello": "world"}}, {"code": {"$scope": {}, "$code": "function x() { return 1; }"}}, {"bin": {"$type": "80", "$binary": "AQIDBA=="}}]')
+   [{u'foo': [1, 2]}, {u'bar': {u'hello': u'world'}}, {u'code': Code('function x() { return 1; }', {})}, {u'bin': Binary('...', 128)}]
 
 Example usage (with a :class:`~bson.json_util.JSONOptions` given):
 
@@ -54,7 +55,7 @@ Example usage (with a :class:`~bson.json_util.JSONOptions` given):
    ...        {'code': Code("function x() { return 1; }")},
    ...        {'bin': Binary(b"\x01\x02\x03\x04")}],
    ...       json_options=CANONICAL_JSON_OPTIONS)
-   '[{"foo": [{"$numberInt": "1"}, {"$numberInt": "2"}]}, {"bar": {"hello": "world"}}, {"code": {"$code": "function x() { return 1; }"}}, {"bin": {"$binary": "AQIDBA==", "$type": "00"}}]'
+   '[{"foo": [{"$numberInt": "1"}, {"$numberInt": "2"}]}, {"bar": {"hello": "world"}}, {"code": {"$code": "function x() { return 1; }"}}, {"bin": {"$binary": {"base64": "AQIDBA==", "subType": "00"}}}]'
 
 Alternatively, you can manually pass the `default` to :func:`json.dumps`.
 It won't handle :class:`~bson.binary.Binary` and :class:`~bson.code.Code`
@@ -92,7 +93,6 @@ but it will be faster as there is less recursion.
 """
 
 import base64
-import collections
 import datetime
 import math
 import re
@@ -114,7 +114,7 @@ else:
 from pymongo.errors import ConfigurationError
 
 import bson
-from bson import EPOCH_AWARE, RE_TYPE, SON
+from bson import EPOCH_AWARE, EPOCH_NAIVE, RE_TYPE, SON
 from bson.binary import (Binary, JAVA_LEGACY, CSHARP_LEGACY, OLD_UUID_SUBTYPE,
                          UUID_SUBTYPE)
 from bson.code import Code
@@ -125,7 +125,8 @@ from bson.int64 import Int64
 from bson.max_key import MaxKey
 from bson.min_key import MinKey
 from bson.objectid import ObjectId
-from bson.py3compat import PY3, iteritems, integer_types, string_type, text_type
+from bson.py3compat import (PY3, iteritems, integer_types, string_type,
+                            text_type)
 from bson.regex import Regex
 from bson.timestamp import Timestamp
 from bson.tz_util import utc
@@ -180,6 +181,30 @@ class DatetimeRepresentation:
     """
 
 
+class JSONMode:
+    LEGACY = 0
+    """Legacy Extended JSON representation.
+
+    .. versionadded:: 3.5
+    """
+
+    RELAXED = 1
+    """Relaxed Extended JSON representation.
+
+    .. seealso:: The specification for Relaxed `Extended JSON`_.
+
+    .. versionadded:: 3.5
+    """
+
+    CANONICAL = 2
+    """Canonical Extended JSON representation.
+
+    .. seealso:: The specification for Canonical `Extended JSON`_.
+
+    .. versionadded:: 3.5
+    """
+
+
 class JSONOptions(CodecOptions):
     """Encapsulates JSON options for :func:`dumps` and :func:`loads`.
 
@@ -198,11 +223,8 @@ class JSONOptions(CodecOptions):
       - `strict_uuid`: If ``True``, :class:`uuid.UUID` object are encoded to
         MongoDB Extended JSON's *Strict mode* type `Binary`. Otherwise it
         will be encoded as ``'{"$uuid": "<hex>" }'``. Defaults to ``False``.
-      - `canonical_extended_json`: If ``True``, use Canonical Extended JSON
-        representations for all BSON values. This option implies
-        ``strict_number_long=True``,
-        ``datetime_representation=DatetimeRepresentation.NUMBERLONG``, and
-        ``strict_uuid=True``.
+      - `json_mode`: The JSON representation to use for encoding and
+        decoding Extended JSON`. Defaults to :const:`~JSONMode.LEGACY`.
       - `document_class`: BSON documents returned by :func:`loads` will be
         decoded to an instance of this class. Must be a subclass of
         :class:`collections.MutableMapping`. Defaults to :class:`dict`.
@@ -219,19 +241,21 @@ class JSONOptions(CodecOptions):
       - `args`: arguments to :class:`~bson.codec_options.CodecOptions`
       - `kwargs`: arguments to :class:`~bson.codec_options.CodecOptions`
 
+    .. seealso:: The `Extended JSON`_ specification.
+
     .. seealso:: The documentation for `MongoDB Extended JSON
        <http://www.mongodb.org/display/DOCS/Mongo+Extended+JSON>`_.
 
     .. versionadded:: 3.4
 
     .. versionchanged:: 3.5
-       Accepts the optional parameter `canonical_extended_json`.
+       Accepts the optional parameter `json_mode`.
 
     """
 
     def __new__(cls, strict_number_long=False,
                 datetime_representation=DatetimeRepresentation.LEGACY,
-                strict_uuid=False, canonical_extended_json=False,
+                strict_uuid=False, json_mode=JSONMode.LEGACY,
                 *args, **kwargs):
         kwargs["tz_aware"] = kwargs.get("tz_aware", True)
         if kwargs["tz_aware"]:
@@ -240,7 +264,7 @@ class JSONOptions(CodecOptions):
                                            DatetimeRepresentation.NUMBERLONG,
                                            DatetimeRepresentation.ISO8601):
             raise ConfigurationError(
-                "JSONOptions.datetime_representation must be one of LEGACY,"
+                "JSONOptions.datetime_representation must be one of LEGACY, "
                 "NUMBERLONG, or ISO8601 from DatetimeRepresentation.")
         self = super(JSONOptions, cls).__new__(cls, *args, **kwargs)
         if not _HAS_OBJECT_PAIRS_HOOK and self.document_class != dict:
@@ -248,8 +272,18 @@ class JSONOptions(CodecOptions):
                 "Support for JSONOptions.document_class on Python 2.6 "
                 "requires simplejson "
                 "(https://pypi.python.org/pypi/simplejson) to be installed.")
-        self.canonical_extended_json = canonical_extended_json
-        if self.canonical_extended_json:
+        if json_mode not in (JSONMode.LEGACY,
+                             JSONMode.RELAXED,
+                             JSONMode.CANONICAL):
+            raise ConfigurationError(
+                "JSONOptions.json_mode must be one of LEGACY, RELAXED, "
+                "or CANONICAL from JSONMode.")
+        self.json_mode = json_mode
+        if self.json_mode == JSONMode.RELAXED:
+            self.strict_number_long = False
+            self.datetime_representation = DatetimeRepresentation.ISO8601
+            self.strict_uuid = True
+        elif self.json_mode == JSONMode.CANONICAL:
             self.strict_number_long = True
             self.datetime_representation = DatetimeRepresentation.NUMBERLONG
             self.strict_uuid = True
@@ -262,33 +296,53 @@ class JSONOptions(CodecOptions):
     def _arguments_repr(self):
         return ('strict_number_long=%r, '
                 'datetime_representation=%r, '
-                'strict_uuid=%r, canonical_extended_json=%r, %s' % (
+                'strict_uuid=%r, json_mode=%r, %s' % (
                     self.strict_number_long,
                     self.datetime_representation,
                     self.strict_uuid,
-                    self.canonical_extended_json,
+                    self.json_mode,
                     super(JSONOptions, self)._arguments_repr()))
 
 
-DEFAULT_JSON_OPTIONS = JSONOptions()
+LEGACY_JSON_OPTIONS = JSONOptions(json_mode=JSONMode.LEGACY)
+""":class:`JSONOptions` for encoding to PyMongo's legacy JSON format.
+
+.. versionadded:: 3.5
+"""
+
+DEFAULT_JSON_OPTIONS = LEGACY_JSON_OPTIONS
 """The default :class:`JSONOptions` for JSON encoding/decoding.
 
+The same as :const:`LEGACY_JSON_OPTIONS`. This will change to
+:const:`RELAXED_JSON_OPTIONS` in a future release.
+
 .. versionadded:: 3.4
+"""
+
+CANONICAL_JSON_OPTIONS = JSONOptions(json_mode=JSONMode.CANONICAL)
+""":class:`JSONOptions` for Canonical `Extended JSON`_.
+
+.. versionadded:: 3.5
+"""
+
+RELAXED_JSON_OPTIONS = JSONOptions(json_mode=JSONMode.RELAXED)
+""":class:`JSONOptions` for Relaxed `Extended JSON`_.
+
+.. versionadded:: 3.5
 """
 
 STRICT_JSON_OPTIONS = JSONOptions(
     strict_number_long=True,
     datetime_representation=DatetimeRepresentation.ISO8601,
     strict_uuid=True)
-""":class:`JSONOptions` for MongoDB Extended JSON's *Strict mode* encoding.
+"""DEPRECATED :class:`JSONOptions` for MongoDB Extended JSON's *Strict mode*
+encoding.
 
 .. versionadded:: 3.4
-"""
 
-CANONICAL_JSON_OPTIONS = JSONOptions(canonical_extended_json=True)
-""":class:`JSONOptions` for `canonical extended JSON`_.
-
-.. versionadded:: 3.5
+.. versionchanged:: 3.5
+   Deprecated. Use :const:`RELAXED_JSON_OPTIONS` or
+   :const:`CANONICAL_JSON_OPTIONS` instead.
 """
 
 
@@ -332,10 +386,7 @@ def loads(s, *args, **kwargs):
         kwargs["object_pairs_hook"] = lambda pairs: object_pairs_hook(
             pairs, json_options)
     else:
-        kwargs["object_hook"] = lambda obj: (
-            canonical_object_hook(obj, json_options)
-            if json_options.canonical_extended_json
-            else object_hook(obj, json_options))
+        kwargs["object_hook"] = lambda obj: object_hook(obj, json_options)
     return json.loads(s, *args, **kwargs)
 
 
@@ -355,54 +406,73 @@ def _json_convert(obj, json_options=DEFAULT_JSON_OPTIONS):
 
 
 def object_pairs_hook(pairs, json_options=DEFAULT_JSON_OPTIONS):
-    document = json_options.document_class(pairs)
-    if json_options.canonical_extended_json:
-        return canonical_object_hook(document, json_options)
-    return object_hook(document, json_options)
+    return object_hook(json_options.document_class(pairs), json_options)
 
 
 def object_hook(dct, json_options=DEFAULT_JSON_OPTIONS):
     if "$oid" in dct:
-        return ObjectId(str(dct["$oid"]))
+        return _parse_canonical_oid(dct)
     if "$ref" in dct:
-        return DBRef(dct["$ref"], dct["$id"], dct.get("$db", None))
+        return _parse_canonical_dbref(dct)
     if "$date" in dct:
-        return _get_date(dct, json_options)
+        return _parse_canonical_datetime(dct, json_options)
     if "$regex" in dct:
-        flags = 0
-        # PyMongo always adds $options but some other tools may not.
-        for opt in dct.get("$options", ""):
-            flags |= _RE_OPT_TABLE.get(opt, 0)
-        return Regex(dct["$regex"], flags)
+        return _parse_legacy_regex(dct)
     if "$minKey" in dct:
-        return MinKey()
+        return _parse_canonical_minkey(dct)
     if "$maxKey" in dct:
-        return MaxKey()
+        return _parse_canonical_maxkey(dct)
     if "$binary" in dct:
-        return _get_binary(dct, json_options)
+        if "$type" in dct:
+            return _parse_legacy_binary(dct, json_options)
+        else:
+            return _parse_canonical_binary(dct, json_options)
     if "$code" in dct:
-        return Code(dct["$code"], dct.get("$scope"))
+        return _parse_canonical_code(dct)
     if "$uuid" in dct:
-        return uuid.UUID(dct["$uuid"])
+        return _parse_legacy_uuid(dct)
     if "$undefined" in dct:
         return None
     if "$numberLong" in dct:
-        return Int64(dct["$numberLong"])
+        return _parse_canonical_int64(dct)
     if "$timestamp" in dct:
         tsp = dct["$timestamp"]
         return Timestamp(tsp["t"], tsp["i"])
     if "$numberDecimal" in dct:
-        return Decimal128(dct["$numberDecimal"])
+        return _parse_canonical_decimal128(dct)
+    if "$dbPointer" in dct:
+        return _parse_canonical_dbpointer(dct)
+    if "$regularExpression" in dct:
+        return _parse_canonical_regex(dct)
+    if "$symbol" in dct:
+        return _parse_canonical_symbol(dct)
+    if "$numberInt" in dct:
+        return _parse_canonical_int32(dct)
+    if "$numberDouble" in dct:
+        return _parse_canonical_double(dct)
     return dct
 
 
-def _get_binary(doc, json_options):
-    if isinstance(doc["$type"], int):
-        doc["$type"] = "%02x" % doc["$type"]
-    subtype = int(doc["$type"], 16)
-    if subtype >= 0xffffff80:  # Handle mongoexport values
-        subtype = int(doc["$type"][6:], 16)
-    data = base64.b64decode(doc["$binary"].encode())
+def _parse_legacy_regex(doc):
+    pattern = doc["$regex"]
+    # Check if this is the $regex query operator.
+    if isinstance(pattern, Regex):
+        return doc
+    flags = 0
+    # PyMongo always adds $options but some other tools may not.
+    for opt in doc.get("$options", ""):
+        flags |= _RE_OPT_TABLE.get(opt, 0)
+    return Regex(pattern, flags)
+
+
+def _parse_legacy_uuid(doc):
+    """Decode a JSON legacy $uuid to Python UUID."""
+    if len(doc) != 1:
+        raise TypeError('Bad $uuid, extra field(s): %s' % (doc,))
+    return uuid.UUID(doc["$uuid"])
+
+
+def _binary_or_uuid(data, subtype, json_options):
     # special handling for UUID
     if subtype == OLD_UUID_SUBTYPE:
         if json_options.uuid_representation == CSHARP_LEGACY:
@@ -417,8 +487,38 @@ def _get_binary(doc, json_options):
     return Binary(data, subtype)
 
 
-def _get_date(doc, json_options):
+def _parse_legacy_binary(doc, json_options):
+    if isinstance(doc["$type"], int):
+        doc["$type"] = "%02x" % doc["$type"]
+    subtype = int(doc["$type"], 16)
+    if subtype >= 0xffffff80:  # Handle mongoexport values
+        subtype = int(doc["$type"][6:], 16)
+    data = base64.b64decode(doc["$binary"].encode())
+    return _binary_or_uuid(data, subtype, json_options)
+
+
+def _parse_canonical_binary(doc, json_options):
+    binary = doc["$binary"]
+    b64 = binary["base64"]
+    subtype = binary["subType"]
+    if not isinstance(b64, string_type):
+        raise TypeError('$binary base64 must be a string: %s' % (doc,))
+    if not isinstance(subtype, string_type) or len(subtype) > 2:
+        raise TypeError('$binary subType must be a string at most 2 '
+                        'characters: %s' % (doc,))
+    if len(binary) != 2:
+        raise TypeError('$binary must include only "base64" and "subType" '
+                        'components: %s' % (doc,))
+
+    data = base64.b64decode(b64.encode())
+    return _binary_or_uuid(data, int(subtype, 16), json_options)
+
+
+def _parse_canonical_datetime(doc, json_options):
+    """Decode a JSON datetime to python datetime.datetime."""
     dtm = doc["$date"]
+    if len(doc) != 1:
+        raise TypeError('Bad $date, extra field(s): %s' % (doc,))
     # mongoexport 2.6 and newer
     if isinstance(dtm, string_type):
         # Parse offset
@@ -441,8 +541,16 @@ def _get_date(doc, json_options):
             dt = dtm
             offset = ''
 
+        # Parse the optional factional seconds portion.
+        dot_index = dt.rfind('.')
+        microsecond = 0
+        if dot_index != -1:
+            microsecond = int(float(dt[dot_index:]) * 1000000)
+            dt = dt[:dot_index]
+
         aware = datetime.datetime.strptime(
-            dt, "%Y-%m-%dT%H:%M:%S.%f").replace(tzinfo=utc)
+            dt, "%Y-%m-%dT%H:%M:%S").replace(microsecond=microsecond,
+                                             tzinfo=utc)
 
         if offset and offset != 'Z':
             if len(offset) == 6:
@@ -462,68 +570,139 @@ def _get_date(doc, json_options):
             return aware
         else:
             return aware.replace(tzinfo=None)
-    # mongoexport 2.6 and newer, time before the epoch (SERVER-15275)
-    elif isinstance(dtm, collections.Mapping):
-        millis = int(dtm["$numberLong"])
-    # mongoexport before 2.6
-    else:
-        millis = int(dtm)
-    return bson._millis_to_datetime(millis, json_options)
+    return bson._millis_to_datetime(int(dtm), json_options)
 
 
-def _get_dbpointer(doc, json_options):
-    dbref = doc['$dbPointer']
-    if isinstance(dbref, DBRef):
-        # DBPointer must not contain $db in its value.
-        if dbref.database is None:
-            return dbref
-        # Otherwise, this is just a regular document.
-        return json_options.document_class(
-            [('$dbPointer', json_options.document_class(dbref.as_doc()))])
-    return doc
+def _parse_canonical_oid(doc):
+    """Decode a JSON ObjectId to bson.objectid.ObjectId."""
+    if len(doc) != 1:
+        raise TypeError('Bad $oid, extra field(s): %s' % (doc,))
+    return ObjectId(doc['$oid'])
 
 
-_CANONICAL_JSON_TABLE = {
-    frozenset(['$oid']): lambda d, _: ObjectId(d['$oid']),
-    frozenset(['$numberDecimal']): lambda d, _: Decimal128(d['$numberDecimal']),
-    frozenset(['$symbol']): lambda d, _: text_type(d['$symbol']),
-    frozenset(['$numberInt']): lambda d, _: int(d['$numberInt']),
-    frozenset(['$numberDouble']): lambda d, _: float(d['$numberDouble']),
-    frozenset(['$numberLong']): lambda d, _: Int64(d['$numberLong']),
-    frozenset(['$date']): _get_date,
-    frozenset(['$minKey']): lambda dummy0, dummy1: MinKey(),
-    frozenset(['$maxKey']): lambda dummy0, dummy1: MaxKey(),
-    frozenset(['$undefined']): lambda dummy0, dummy1: None,
-    frozenset(['$dbPointer']): _get_dbpointer,
-    frozenset(['$ref', '$id']): lambda d, _: DBRef(
-        d.pop('$ref'), d.pop('$id'), **d),
-    frozenset(['$ref', '$id', '$db']): lambda d, _: DBRef(
-        d.pop('$ref'), d.pop('$id'), d.pop('$db'), **d),
-    frozenset(['$regex', '$options']): lambda d, _: Regex(
-        d['$regex'], d['$options']),
-    frozenset(['$binary', '$type']): _get_binary,
-    frozenset(['$code']): lambda d, _: Code(d['$code']),
-    frozenset(['$code', '$scope']): lambda d, _: Code(
-        d['$code'], d['$scope']),
-    frozenset(['$timestamp']): lambda d, _: Timestamp(
-        int(d['$timestamp']) >> 32, int(d['$timestamp']) & 0xffffffff)
-}
+def _parse_canonical_symbol(doc):
+    """Decode a JSON symbol to Python string."""
+    symbol = doc['$symbol']
+    if len(doc) != 1:
+        raise TypeError('Bad $symbol, extra field(s): %s' % (doc,))
+    return text_type(symbol)
 
 
-def canonical_object_hook(dct, json_options=CANONICAL_JSON_OPTIONS):
-    keyset = frozenset(key for key in dct if key.startswith('$'))
-    converter = _CANONICAL_JSON_TABLE.get(keyset)
-    if converter:
-        return converter(dct, json_options)
-    elif '$ref' in dct and '$id' in dct:
-        # DBRef may contain other keys that don't start with $.
-        if keyset - _DBREF_KEYS:
+def _parse_canonical_code(doc):
+    """Decode a JSON code to bson.code.Code."""
+    for key in doc:
+        if key not in ('$code', '$scope'):
+            raise TypeError('Bad $code, extra field(s): %s' % (doc,))
+    return Code(doc['$code'], scope=doc.get('$scope'))
+
+
+def _parse_canonical_regex(doc):
+    """Decode a JSON regex to bson.regex.Regex."""
+    regex = doc['$regularExpression']
+    if len(doc) != 1:
+        raise TypeError('Bad $regularExpression, extra field(s): %s' % (doc,))
+    if len(regex) != 2:
+        raise TypeError('Bad $regularExpression must include only "pattern"'
+                        'and "options" components: %s' % (doc,))
+    return Regex(regex['pattern'], regex['options'])
+
+
+def _parse_canonical_dbref(doc):
+    """Decode a JSON DBRef to bson.dbref.DBRef."""
+    for key in doc:
+        if key.startswith('$') and key not in _DBREF_KEYS:
             # Other keys start with $, so dct cannot be parsed as a DBRef.
-            return dct
-        else:
-            return DBRef(dct.pop('$ref'), dct.pop('$id'),
-                         dct.pop('$db', None), **dct)
-    return dct
+            return doc
+    return DBRef(doc.pop('$ref'), doc.pop('$id'),
+                 database=doc.pop('$db', None), **doc)
+
+
+def _parse_canonical_dbpointer(doc):
+    """Decode a JSON (deprecated) DBPointer to bson.dbref.DBRef."""
+    dbref = doc['$dbPointer']
+    if len(doc) != 1:
+        raise TypeError('Bad $dbPointer, extra field(s): %s' % (doc,))
+    if isinstance(dbref, DBRef):
+        dbref_doc = dbref.as_doc()
+        # DBPointer must not contain $db in its value.
+        if dbref.database is not None:
+            raise TypeError(
+                'Bad $dbPointer, extra field $db: %s' % (dbref_doc,))
+        if not isinstance(dbref.id, ObjectId):
+            raise TypeError(
+                'Bad $dbPointer, $id must be an ObjectId: %s' % (dbref_doc,))
+        if len(dbref_doc) != 2:
+            raise TypeError(
+                'Bad $dbPointer, extra field(s) in DBRef: %s' % (dbref_doc,))
+        return dbref
+    else:
+        raise TypeError('Bad $dbPointer, expected a DBRef: %s' % (doc,))
+
+
+def _parse_canonical_int32(doc):
+    """Decode a JSON int32 to python int."""
+    i_str = doc['$numberInt']
+    if len(doc) != 1:
+        raise TypeError('Bad $numberInt, extra field(s): %s' % (doc,))
+    if not isinstance(i_str, string_type):
+        raise TypeError('$numberInt must be string: %s' % (doc,))
+    return int(i_str)
+
+
+def _parse_canonical_int64(doc):
+    """Decode a JSON int64 to bson.int64.Int64."""
+    l_str = doc['$numberLong']
+    if len(doc) != 1:
+        raise TypeError('Bad $numberLong, extra field(s): %s' % (doc,))
+    return Int64(l_str)
+
+
+def _parse_canonical_double(doc):
+    """Decode a JSON double to python float."""
+    d_str = doc['$numberDouble']
+    if len(doc) != 1:
+        raise TypeError('Bad $numberDouble, extra field(s): %s' % (doc,))
+    if not isinstance(d_str, string_type):
+        raise TypeError('$numberDouble must be string: %s' % (doc,))
+    return float(d_str)
+
+
+def _parse_canonical_decimal128(doc):
+    """Decode a JSON decimal128 to bson.decimal128.Decimal128."""
+    d_str = doc['$numberDecimal']
+    if len(doc) != 1:
+        raise TypeError('Bad $numberDecimal, extra field(s): %s' % (doc,))
+    if not isinstance(d_str, string_type):
+        raise TypeError('$numberDecimal must be string: %s' % (doc,))
+    return Decimal128(d_str)
+
+
+def _parse_canonical_minkey(doc):
+    """Decode a JSON MinKey to bson.min_key.MinKey."""
+    if doc['$minKey'] is not 1:
+        raise TypeError('$minKey value must be 1: %s' % (doc,))
+    if len(doc) != 1:
+        raise TypeError('Bad $minKey, extra field(s): %s' % (doc,))
+    return MinKey()
+
+
+def _parse_canonical_maxkey(doc):
+    """Decode a JSON MaxKey to bson.max_key.MaxKey."""
+    if doc['$maxKey'] is not 1:
+        raise TypeError('$maxKey value must be 1: %s', (doc,))
+    if len(doc) != 1:
+        raise TypeError('Bad $minKey, extra field(s): %s' % (doc,))
+    return MaxKey()
+
+
+def _encode_binary(data, subtype, json_options):
+    if json_options.json_mode == JSONMode.LEGACY:
+        return SON([
+            ('$binary', base64.b64encode(data).decode()),
+            ('$type', "%02x" % subtype)])
+    return {'$binary': SON([
+        ('base64', base64.b64encode(data).decode()),
+        ('subType', "%02x" % subtype)])}
 
 
 def default(obj, json_options=DEFAULT_JSON_OPTIONS):
@@ -544,10 +723,10 @@ def default(obj, json_options=DEFAULT_JSON_OPTIONS):
                     tz_string = 'Z'
                 else:
                     tz_string = obj.strftime('%z')
-                return {"$date": "%s.%03d%s" % (
-                    obj.strftime("%Y-%m-%dT%H:%M:%S"),
-                    int(obj.microsecond / 1000),
-                    tz_string)}
+                millis = int(obj.microsecond / 1000)
+                fracsecs = ".%03d" % (millis,) if millis else ""
+                return {"$date": "%s%s%s" % (
+                    obj.strftime("%Y-%m-%dT%H:%M:%S"), fracsecs, tz_string)}
 
         millis = bson._datetime_to_millis(obj)
         if (json_options.datetime_representation ==
@@ -574,29 +753,26 @@ def default(obj, json_options=DEFAULT_JSON_OPTIONS):
             pattern = obj.pattern
         else:
             pattern = obj.pattern.decode('utf-8')
-        return SON([("$regex", pattern), ("$options", flags)])
+        if json_options.json_mode == JSONMode.LEGACY:
+            return SON([("$regex", pattern), ("$options", flags)])
+        return {'$regularExpression': SON([("pattern", pattern),
+                                           ("options", flags)])}
     if isinstance(obj, MinKey):
         return {"$minKey": 1}
     if isinstance(obj, MaxKey):
         return {"$maxKey": 1}
     if isinstance(obj, Timestamp):
-        if json_options.canonical_extended_json:
-            return {'$timestamp': str((obj.time << 32) + obj.inc)}
         return {"$timestamp": SON([("t", obj.time), ("i", obj.inc)])}
     if isinstance(obj, Code):
         if obj.scope is None:
-            return SON([('$code', str(obj))])
+            return {'$code': str(obj)}
         return SON([
             ('$code', str(obj)),
             ('$scope', _json_convert(obj.scope, json_options))])
     if isinstance(obj, Binary):
-        return SON([
-            ('$binary', base64.b64encode(obj).decode()),
-            ('$type', "%02x" % obj.subtype)])
+        return _encode_binary(obj, obj.subtype, json_options)
     if PY3 and isinstance(obj, bytes):
-        return SON([
-            ('$binary', base64.b64encode(obj).decode()),
-            ('$type', "00")])
+        return _encode_binary(obj, 0, json_options)
     if isinstance(obj, uuid.UUID):
         if json_options.strict_uuid:
             data = obj.bytes
@@ -607,28 +783,27 @@ def default(obj, json_options=DEFAULT_JSON_OPTIONS):
                 data = data[7::-1] + data[:7:-1]
             elif json_options.uuid_representation == UUID_SUBTYPE:
                 subtype = UUID_SUBTYPE
-            return SON([
-                ('$binary', base64.b64encode(data).decode()),
-                ('$type', "%02x" % subtype)])
+            return _encode_binary(data, subtype, json_options)
         else:
             return {"$uuid": obj.hex}
     if isinstance(obj, Decimal128):
         return {"$numberDecimal": str(obj)}
     if isinstance(obj, bool):
         return obj
-    if json_options.canonical_extended_json and isinstance(obj, integer_types):
+    if (json_options.json_mode == JSONMode.CANONICAL and
+            isinstance(obj, integer_types)):
         if -2 ** 31 <= obj < 2 ** 31:
             return {'$numberInt': text_type(obj)}
         return {'$numberLong': text_type(obj)}
-    if json_options.canonical_extended_json and isinstance(obj, float):
+    if json_options.json_mode != JSONMode.LEGACY and isinstance(obj, float):
         if math.isnan(obj):
-            representation = 'NaN'
+            return {'$numberDouble': 'NaN'}
         elif math.isinf(obj):
             representation = 'Infinity' if obj > 0 else '-Infinity'
-        else:
+            return {'$numberDouble': representation}
+        elif json_options.json_mode == JSONMode.CANONICAL:
             # repr() will return the shortest string guaranteed to produce the
             # original value, when float() is called on it. str produces a
             # shorter string in Python 2.
-            representation = text_type(repr(obj))
-        return {'$numberDouble': representation}
+            return {'$numberDouble': text_type(repr(obj))}
     raise TypeError("%r is not JSON serializable" % obj)